refactor: move product existence check to ProductService.deleteById returning boolean

+++ src/main/java/com/andres/course/agy/springboot/springmvc/app/services/ProductServiceImpl.java
@@ -31,7 +31,11 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
